Fix unit test fails
